requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==43.0.0
defusedxml==0.7.1
Django==5.0.7
django-allauth==0.61.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.13
django-filter==24.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
oauthlib==3.2.2
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
